refactor(draft-editor): rename misspelled file
