+++ packages/draft-editor/src/draft-js/editor/block-renderer-fn.tsx
